+++ src/edu/gatech/lbs/sim/tracegenerator/paramdistribution/IParamDistribution.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2009, Georgia Tech Research Corporation
+// Copyright (c) 2012, Georgia Tech Research Corporation
 // Authors:
 //   Peter Pesti (pesti@gatech.edu)
 //
@@ -8,6 +8,6 @@ import edu.gatech.lbs.core.vector.IVector;
 
 public interface IParamDistribution {
 
-  public double getNextValue(IVector location);
+  public int getNextValue(IVector location);
 
 }
